passthroughfs: fix a race window between do_readdir() and lseek()
The do_readdir() function is implemented by:
1) lseek
2) getdents64
So it's a must to keep lseek and getdents64 are called in an atomic
context. And this is the only case needing exclusive access to the
underlying file descriptor.

The solution is to protect do_readdir() and lseek() by a lock.
Currently design is aimed to implement the protection, but failed
its goal. The call to libc::lseek() in function lseek() is actually
not protected by any lock. So fix the race window.

Because this is the only case needing exclusive access, so it's
possible to relax other accesses without acquiring any lock.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>