passthrufs: enable killpriv_v2 support for sync mode
When kernel supports HANDLING_KILL_PRIV_V2, we can do kill_priv in the
file system so that kernel doesn't have to query xattr on each write and
it helps to improve performance a lot.

For async mode, we cannot simply toggle thread caps before and after
doing create. So it is skipped at the moment.

Signed-off-by: Peng Tao <bergwolf@hyper.sh>