passthroughfs: Make InodeAltKey an enum
This patch is syntactical preparation to make the next patch a bit
easier.

`InodeAltKey` is made an enum, with only an `Ids` variant right now that
contains the existing st_ino/st_dev/mnt_id combination.  The next patch
will add a `Handle` variant so a file handle can be added as an
alternate key for an inode.

Signed-off-by: Max Reitz <mreitz@redhat.com>
Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>