passthrough: open file for reading if writeback cache is enabled
When writeback caching is enabled, the kernel may send read requests
even if the userspace program opened the file write-only. So we need to
ensure that we have opened the file for reading as well as writing.

But we only did this when handling OPEN request, CREATE request with
O_WRONLY still misses read permission. This results in EBADF error when
writing such file from client side.

Fix it by introducing get_writeback_open_flags() function, which ensures
read permission is granted, and call this helper function in both open
and create code path.

Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>