passthroughfs: convert MultiKeyBTreeMap to InodeStore for InodeMap
The implementation of MultiKeyBTreeMap is for generic usage and may not
be efficient in passthroughfs, e.g. it has a vector in alt key and when
InodeMap grows, it triggers many memory re-allocations, and hits
performance.

So introduce application-specific InodeStore struct to replace the
generic MultiKeyBTreeMap. This follows virtiofsd-rs commit 4d27d3ec772f
("passthrough: Use application-specific inode storage")

Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>